cash.xml
This project didn't stop (as the commits suggest) - Instead I created a new temporary project to seperately rework the loose change aspect with a GUI, using an xml file (because I hadn't used xml before and felt the need to try it). I've since been using that to track my cash; But now I'm returning to this project and merging the GUIs back in to the overall application.

Committing now (post-merge) before any real work begins on the rest of the features.