Dates work
Been working on detecting which bills have already been paid (between last payday and today), and which have still to be paid (those with due dates before next payday).

Also now have it working so that you could leave the app unused for long periods of time and it would update the payment dates etc to catch up.

For testing purposes it does not update the csv files yet.