Been tidying up this disaster
This project was such an awful hacked-together mess that I actually had to create an account with Trello and try to map my ideas to it and stick to each task bit by bit.

Still work to be done but a decent amount of refactoring and spring cleaning has been done. At least from now it should be fairly clean to build on top of