Added feature to display how many bank bags full of each coin can be withdrawn
The code is disgusting and may well be refactored in future, but for now it does the job.

Continuation of this feature would include adding to the withdraw shrapnel section of the program an option to withdraw bags rather than making the user count how many coins they are taking out for banking purposes.