Fixed / Updated Income Tax & National Insurance
I've been giving some thought to budgets today and decided to revisit this project. In doing so I discovered that the taxation rates are changing (positively!) next month, so I've updated the income tax calculations to deal with that, as well as getting the national insurance portion working at last.

Now I just need a real salary to use the project with... :sweat_smile: