Currency Conversion
Updated my bills csv to reflect my move from GoDaddy hosting to Heroku, which is paid in USD. My current plan has a monthly fee of $25, so I've set up a free account with CurrencyLayer to get 1000 free requests per month with hourly updates.

At this point I am only using it to convert from $ to £; However it would be very simple to add further capabilities to it the function should I need to reuse it elsewhere.