Edit Bills
Functionality for editing bills is now in.

I've not made adding a new bill possible yet (although an empty row does appear on the form, but does nothing), but making changes to bills is now possible.

Note that much like the previous commit, this section needs validation and some tidying up.

Unsuccessful efforts were made to make the bills list scrollable, so this is still to be done.

I made a duplicate of the BillsPrototype.csv file to maintain the bills should anything go wrong (and it did, multiple times :sweat_smile:)

I should also note that using this will re-order your .csv file, as the bills list is sorted based on payment dates after being read in. Returning them to the file will result in them taking the new form